+++ docs/Changelog.md
@@ -30,6 +30,7 @@
   [#5734](https://github.com/TouK/nussknacker/pull/5734) [#5653](https://github.com/TouK/nussknacker/pull/5653) [#5744](https://github.com/TouK/nussknacker/pull/5744) 
   [#5757](https://github.com/TouK/nussknacker/pull/5757) [#5777](https://github.com/TouK/nussknacker/pull/5777) [#5825](https://github.com/TouK/nussknacker/pull/5825)
   [#5896](https://github.com/TouK/nussknacker/pull/5896) Experimental Batch processing mode support
+* [#6031](https://github.com/TouK/nussknacker/pull/6031) Fixed duplicated components on the Components tab
 
 1.14.0 (21 Mar 2024)
 -------------------------